requirements.txt
pynacl==1.5.0
pycryptodome==3.20.0
python-dotenv==1.0.1